refactor: CourseTableをコンポーネント化
