+++ frontend/src/features/course/index.ts
@@ -1,4 +1,5 @@
 export * from './components/CreateCourseButton'
 export * from './components/UpdateCourseButton'
+export * from './components/CourseTable'
 export * from './hooks/useCourse'
 export * from './types'
